requirements.txt
streamlit
pymupdf
python-docx
pandas
spacy
